resnet build
added htmls too

+++ utils.py
@@ -88,15 +88,15 @@ def load_3d_batch(filename):
     X = np.array(datadict['pixels'])
     Y = datadict['emotion']
     if filename == "train":
-        X = X.reshape(28709, 48, 48).astype("int")
+        X = X.reshape(28709, 48, 48).astype("float")
         X = X[:, :, :, None] * np.ones(3, dtype=int)[None, None, None, :]
         Y = np.array(Y)
     elif filename == "publictest":
-        X = X.reshape(3589, 48, 48).astype("int")
+        X = X.reshape(3589, 48, 48).astype("float")
         X = X[:, :, :, None] * np.ones(3, dtype=int)[None, None, None, :]
         Y = np.array(Y)
     elif filename == "privatetest":
-        X = X.reshape(3589, 48, 48).astype("int")
+        X = X.reshape(3589, 48, 48).astype("float")
         X = X[:, :, :, None] * np.ones(3, dtype=int)[None, None, None, :]
         Y = np.array(Y)
   return X, Y
